+++ kubernetes-client/src/test/java/io/fabric8/kubernetes/client/internal/okhttp/OkHttpClientFactoryTest.java
@@ -18,6 +18,8 @@ package io.fabric8.kubernetes.client.internal.okhttp;
 
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.ConfigBuilder;
+import io.fabric8.kubernetes.client.okhttp.OkHttpClientFactory;
+import io.fabric8.kubernetes.client.okhttp.OkHttpClientImpl;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;