+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/ExecExampleWithTerminalSize.java
@@ -71,7 +71,7 @@ public class ExecExampleWithTerminalSize {
         }
 
         @Override
-        public void onFailure(Throwable t) {
+        public void onFailure(Throwable t, Response failureResponse) {
             System.err.println("shell barfed");
         }