+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/ExecLoopExample.java
@@ -60,7 +60,7 @@ public class ExecLoopExample {
           }
 
           @Override
-          public void onFailure(Throwable t) {
+          public void onFailure(Throwable t, Response failureResponse) {
             latch.countDown();
           }