+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/kubectl/equivalents/PodExecEquivalent.java
@@ -64,7 +64,7 @@ public class PodExecEquivalent {
     }
 
     @Override
-    public void onFailure(Throwable throwable) {
+    public void onFailure(Throwable t, Response failureResponse) {
       logger.info("Some error encountered");
       execLatch.countDown();
     }