+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 * Refactoring #3547: due to an abstraction layer added over okHttp, the following api changes were made:
     * OperationContext withOkHttpClient was removed, it should be needed to be directly called
     * PatchType.getMediaType was replaced with PatchType.getContentType
-    * ExecListener no longer passes the okhttp3.Response.
+    * ExecListener no longer passes the okhttp3.Response to onOpen.  onFailure will pass a simplified ExecListener.Response when possible.
     * okhttp3.TlsVersions has been replaced by io.fabric8.kubernetes.client.http.TlsVersion
     * HttpClientUtils.createHttpClient(config, additionalConfig) has been deprecated and now returns an OkHttpClientImpl